+++ public/javascripts/LoginUtility.js
@@ -70,16 +70,7 @@ console.log("shouldve made it this far");
     else
     return {fuck:"you"};
 }
-function makeid()
-{
-    var text = "";
-    var possible = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
 
-    for( var i=0; i < 30; i++ )
-        text += possible.charAt(Math.floor(Math.random() * possible.length));
-
-    return text;
-}
 
 function isValidLogin(username,password) {
